+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All changes from version 1.1.1 will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.4.3] - Not Release Yet
+- add simple PKI utils
+- add default syslog sink component
+- add ssl support for syslog semantics
+
 ## [2.4.3] - 2021-12-07
 - add custom delimiter opt to http source impl
 - update deps,